Support block gap created by fast sync (#5703)
This is part 2 of
https://github.com/paritytech/polkadot-sdk/issues/5406#issuecomment-2325064863,
properly handling the block gap generated during fast sync.

Although #5406 remains unresolved due to the known issues in #5663, I
decided to open up this PR earlier than later to speed up the overall
progress. I've tested the fast sync locally with this PR, and it appears
to be functioning well. (I was doing a fast sync from a discontinued
archive node locally, thus the issue highlighted in
https://github.com/paritytech/polkadot-sdk/pull/5663#discussion_r1752124039
was bypassed exactly.)

Once the edge cases in #5663 are addressed, we can move forward by
removing the body attribute from the LightState block request and
complete the work on #5406. The changes in this PR are incremental, so
reviewing commit by commit should provide the best clarity.

cc @dmitry-markin

---------

Co-authored-by: Bastian Köcher <git@kchr.de>