Implementation of ah-client and rc-client staking pallets (#7582)
Recreation of https://github.com/paritytech/polkadot-sdk/pull/7357 on
top of master. The old PR messes up the git history too much so I am
recreating it from scratch.

This PR is work in progress. It's purpose is to commit initial structure
of `pallet-staking-ah-client` and `pallet-staking-rc-client` to master.
The changes will be polished by a follow up PRs which will be
backported.

Related issues: https://github.com/paritytech/polkadot-sdk/issues/6167
and https://github.com/paritytech/polkadot-sdk/issues/6166

This PR introduces the initial structure for `pallet-ah-client` and
`pallet-rc-client`. These pallets will reside on the relay chain and
AssetHub, respectively, and will manage the interaction between
`pallet-session` on the relay chain and `pallet-staking` on AssetHub.

Both pallets are experimental and not intended for production use.

TODOs:
- [ ] Probably handle
https://github.com/paritytech/polkadot-sdk/issues/6344 here.

---------

Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Ankan <10196091+Ank4n@users.noreply.github.com>