[pallet-broker] add extrinsic to remove a lease (#7026)
# Description

#6929 requests more extrinsics for "managing the network's coretime
allocations without needing to dabble with migration+runtime upgrade or
set/kill storage patterns"

This pull request implements the remove_lease() extrinsic.


## Integration

Downstream projects need to benchmark the weight for the remove_lease()
extrinsic.

## Review Notes

Mentorship is requested to ensure this is implemented correctly.

The lease is removed from state using the TaskId as a key. Is this
sufficient. Does the extrinsic need to do anything else?

---------

Co-authored-by: Jonathan Brown <jbrown@acuity.network>
Co-authored-by: Bastian Köcher <git@kchr.de>
Co-authored-by: command-bot <>
Co-authored-by: Dónal Murray <donalm@seadanda.dev>
Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Dónal Murray <donal.murray@parity.io>