tx/broadcast: Use client info for fetching the best block (#8162)
This PR modifies the tx broadcast logic. Previously, we relied on the
import notification stream to fetch the latest available block. In this
revision, we use the available information of the `client.info()` to
prefix the import notification stream.

The change is needed because chains could be started with `InstantSeal`,
which would not produce blocks until they encounter transactions.

Closes: https://github.com/paritytech/polkadot-sdk/issues/8157

cc @paritytech/subxt-team

---------

Signed-off-by: Alexandru Vasile <alexandru.vasile@parity.io>
Co-authored-by: Bastian Köcher <git@kchr.de>