Add methods in the statement store RPCs to get the statements and not just the statement data. (#8314)
Some statement contain a proof with the signature of the statement, this
proof is useful to assert that the statement comes from the expected
account.

Alternatively we can always add some signature inside the encrypted
data, but this means we sign 2 times and it isn't necessary in some
cases.

What do you think @arkpar ?

---------

Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>