txpool: fix tx removal from unlocks set (#8500)
# Description

Removing a tx subtree means partly removing some txs from the unlocks
set of other txs. This logic is buggy and the PR attempts to fix it.

Closes #8498 

## Integration

N/A

## Review Notes

This doesn't seem to be an important bug. Unit tests for txpool still
pass after the fix, so txpool behavior isn't changing much.

### TODOs

- [x] test with a heavy load test (5 millions txs) - all txs were
validated successfully
- [x] added a unit test

---------

Signed-off-by: Iulian Barbu <iulian.barbu@parity.io>
Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>