fix: :bug: use `MaxKeys` from `pallet-im-online`'s Config trait instead of hardcoded one in benchmarks (#9325)
This PR is a simple fix for issue #9324, by making the benchmarks of
`pallet-im-online` linear up to `pallet_im_online::Config::MaxKeys`
instead of the hardcoded constant `MAX_KEYS = 1000`.

This should allow any runtime that uses `pallet-im-online` with less
than 1000 max keys to be able to benchmark the pallet correctly.