pallet_revive: Enforce storage deposit limit on plain transfer (#9416)
The existential deposit to create a new account is part of the storage
deposit. Hence if the storage deposit limit is too low to create a new
account we fail the transaction. However, this limit was not enforced
for plain transfers. The reason is that we only enforce the limit at the
end of each frame. But for plain transfers (transferring to a non
contract) there is no frame.

This PR fixes the situation by enforcing the limit when transferring the
existential deposit in order to create a new account.

---------

Co-authored-by: PG Herveou <pgherveou@gmail.com>
Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>