store headers and justifications during warp sync (#9424)
Closes https://github.com/paritytech/polkadot-sdk/issues/2738.

Still need to add tests for this - but I think the easiest way might be
after the zombienet tests are converted to Rust, in the warp sync test
maybe we can just request the headers (and justifications?) from
JSON-RPC? Though I'm not sure there is an API for the justifications.
But in any case we can in theory make a P2P justifications request as
well and the node should be able to respond. Let me know if anybody has
some better ideas.

---------

Signed-off-by: sistemd <enntheprogrammer@gmail.com>
Co-authored-by: cmd[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Iulian Barbu <14218860+iulianbarbu@users.noreply.github.com>